# gbvmis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a gbvmis api",
        "title": "SGBV Police API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080/api",
    "basePath": "/",
    "paths": {
...
        "/suspect": {
            "post": {
                "description": "Creates a new suspect entry with photo and fingerprints files.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suspects"
                ],
                "summary": "Create a new suspect record with photo and fingerprints upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "First Name",
                        "name": "first_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Middle Name",
                        "name": "middle_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Last Name",
                        "name": "last_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "DOB in YYYY-MM-DD format",
                        "name": "dob",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Gender",
                        "name": "gender",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Phone Number",
                        "name": "phone_number",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "NIN",
                        "name": "nin",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nationality",
                        "name": "nationality",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Occupation",
                        "name": "occupation",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Created By",
                        "name": "created_by",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Updated By",
                        "name": "updated_by",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Photo file upload",
                        "name": "photo",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Fingerprints file upload",
                        "name": "fingerprints",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created suspect record",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid input or file error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when creating suspect",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/suspect/{id}": {
            "get": {
                "description": "Fetches a suspect record based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suspects"
                ],
                "summary": "Retrieve a single suspect record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Suspect ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suspect retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Suspect not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when retrieving suspect",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "put": {
                "description": "Partially updates fields of an existing suspect, including photo and fingerprints.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suspects"
                ],
                "summary": "Update a suspect record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Suspect ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "First Name",
                        "name": "first_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Middle Name",
                        "name": "middle_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Last Name",
                        "name": "last_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "DOB in YYYY-MM-DD format",
                        "name": "dob",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Gender",
                        "name": "gender",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Phone Number",
                        "name": "phone_number",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "NIN",
                        "name": "nin",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nationality",
                        "name": "nationality",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Occupation",
                        "name": "occupation",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Updated By",
                        "name": "updated_by",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Photo file upload",
                        "name": "photo",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Fingerprints file upload",
                        "name": "fingerprints",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suspect updated successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid input or file error",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Suspect not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when updating suspect",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a suspect record based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suspects"
                ],
                "summary": "Delete a suspect record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Suspect ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suspect deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Suspect not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when deleting suspect",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/suspects": {
            "get": {
                "description": "Fetches all suspect records with pagination support.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suspects"
                ],
                "summary": "Retrieve a paginated list of suspects",
                "responses": {
                    "200": {
                        "description": "Suspects retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve suspects",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/suspects/search": {
            "get": {
                "description": "Retrieves a paginated list of suspects based on search criteria.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Suspects"
                ],
                "summary": "Search for suspects with pagination",
                "responses": {
                    "200": {
                        "description": "Suspects retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve suspects",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
...
    }
}